addfunc: Split JavaScript files #54
